BACKEND/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-admin-kubi
six
requests==2.31.0
firebase-admin==6.4.0

# Celery and task queue
celery==5.3.6
redis==5.0.1
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Testing dependencies
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
coverage==7.3.2